requirements.txt
Django==2.0.7
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
Pillow==7.2.0
PyJWT==1.7.1
pytz==2020.1
